+++ bin/wabbit-check-inifile.py
@@ -3,17 +3,7 @@
 import sys, os
 sys.path.append(os.path.join(os.path.split(__file__)[0], ".."))
 import inifile_tools
-
-
-class bcolors:
-        HEADER = '\033[95m'
-        OKBLUE = '\033[94m'
-        OKGREEN = '\033[92m'
-        WARNING = '\033[93m'
-        FAIL = '\033[91m'
-        ENDC = '\033[0m'
-        BOLD = '\033[1m'
-        UNDERLINE = '\033[4m'
+import bcolors
 
 # fetch jobfile from call
 inifile = sys.argv[1]
