+++ bin/wabbit-time-left.py
@@ -5,19 +5,11 @@ import sys, os
 sys.path.append(os.path.join(os.path.split(__file__)[0], ".."))
 import inifile_tools
 import insect_tools
+import bcolors
 import glob
 import datetime
 import argparse
 
-class bcolors:
-        HEADER = '\033[95m'
-        OKBLUE = '\033[94m'
-        OKGREEN = '\033[92m'
-        WARNING = '\033[93m'
-        FAIL = '\033[91m'
-        ENDC = '\033[0m'
-        BOLD = '\033[1m'
-        UNDERLINE = '\033[4m'
 
 # this is a peculiar oddity for IRENE, she spits out some runtime warnings....
 np.seterr(invalid='ignore')
