+++ inifile_tools.py
@@ -4,15 +4,6 @@ This script contains functions that deal with modfying the ini files
 import numpy as np
 import bcolors
 
-class bcolors:
-        HEADER = '\033[95m'
-        OKBLUE = '\033[94m'
-        OKGREEN = '\033[92m'
-        WARNING = '\033[93m'
-        FAIL = '\033[91m'
-        ENDC = '\033[0m'
-        BOLD = '\033[1m'
-        UNDERLINE = '\033[4m'
 
 #
 def check_parameters_for_stupid_errors( file ):
